Include software as part of petalinux project
This commit seeks to integrate petalinux into the same VCS tree as the
hardware. The petalinux project includes the kernel configuration and
packages the driver's source code into a petalinux application. This
means that the FPGA's entire software stack can be built by simply
running petalinux-build in the software folder.