Development Environment
This commit includes the development environment I use to work on this
project. If the host linux system has docker/podman and distrobox
installed, running the makefile will automatically build a WSL-like
environment with everything you need to build the project, aside from
Vivado and Petalinux. This helps us ensure that people in the future can
still build the project, since the version of the Xilinx tooling it uses
is not compatible with current linux distributions. It also
automatically manages the TFTP server.